+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 * **HTTP Spoofing:** Added `ONLY_WGET_CURL` option in [`monero-bash.conf`](https://github.com/hinto-janaiyo/monero-bash/blob/main/config/monero-bash.conf) to only use randomly selected 2016-2022 versions of `wget/curl` as the [User-Agent.](https://github.com/hinto-janaiyo/monero-bash/blob/main/docs/fake_http_headers/wget_curl) This option and `ONLY_USER_AGENT` will always enabled when sending a non-local RPC call
 * **HTTP Spoofing:** [More referers [5] -> [9]](https://github.com/hinto-janaiyo/monero-bash/blob/main/docs/fake_http_headers/referer)
 * **HTTP Spoofing:** [More language permutations [230] -> [4105]](https://github.com/hinto-janaiyo/monero-bash/blob/main/docs/fake_http_headers/language)
+* **Status:** P2Pool stat calculation is slightly faster
 
 ## Fixes
 * **Package manager:** The GitHub API returns 1-line JSON (sometimes) which causes parsing errors. This long standing issue was the cause of most update/upgrade failures, it is now fixed
 * **HTTP Spoofing:** Only set once when upgrading multiple packages
+* **Status:** Fixed P2Pool [x/hour, x/day] math
 * **Watch:** Spacing is auto-adjusted on terminal size changes
 
 ---